requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
python-multipart==0.0.9
PyJWT==2.10.1
pymongo==4.15.0
pydantic==2.11.9
httpx==0.28.1
requests==2.32.5
python-dotenv==1.1.1
openai==1.107.2
tenacity==9.1.2
pytest==8.4.2
